Move plugins docs to Guide
